requirements.txt
pandas==2.0.3
torch==2.2.0
scikit-learn==1.3.2
gdown==5.1.0
openai-clip==1.0.1
pillow~=10.2.0
numpy~=1.24.4
tqdm~=4.66.1
scipy~=1.10.1
six~=1.16.0
packaging~=23.2
requests~=2.31.0
ftfy~=6.1.3
regex~=2023.12.25
setuptools~=68.2.2
nltk~=3.8.1
spacy~=3.7.2
pyyaml~=6.0.1
transformers=4.37.2